requirements.txt
pandas==2.2.2
boto3==1.35.0
redshift-connector==2.0.918
pyyaml==6.0.2
pytest==8.3.2
